Garantindo que nao exclua livro sem antes excluir o LivroAutor///chave estrangeira

+++ src/main/java/com/biblioteca/biblioteca/services/LivrosService.java
@@ -1,19 +1,25 @@
 package com.biblioteca.biblioteca.services;
 
 import java.util.Optional;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.biblioteca.biblioteca.domain.LivroAutor;
 import com.biblioteca.biblioteca.domain.Livros;
+import com.biblioteca.biblioteca.exceptions.LivrosAutoresException;
 import com.biblioteca.biblioteca.repository.LivrosRepository;
 
 @Service
 public class LivrosService {
     @Autowired
     private LivrosRepository livrosRepository;
+
+    @Autowired
+    private LivroAutorService livroAutorService;
 
     public Page<Livros> getAll(Pageable pageable){
         return this.livrosRepository.findAll(pageable);
@@ -27,7 +33,14 @@ public class LivrosService {
         return this.livrosRepository.save(livro);
     }
 
-    public void deleteBook(Livros livro){
-        this.livrosRepository.delete(livro);
+    public void deleteBook(Livros livro){   
+        List<LivroAutor> livrosAutores = this.livroAutorService.livrosAutores(livro);
+        if (livrosAutores.isEmpty()){
+            this.livrosRepository.delete(livro);
+        }
+        else{
+            throw new LivrosAutoresException();
+        }
+
     }
 }
